Convert the selectors to work against the new schema.
